Restructure project to apps-based architecture
- Move display app files to /apps/display/ directory structure
  - Pages moved to /apps/display/pages/
  - Lib files (layout-optimizer, webrtc) moved to /apps/display/lib/
  - Hooks moved to /apps/display/hooks/
- Update Next.js routes in /src/app/apps/display/ to import from /apps/display/
- Remove duplicate routes (/src/app/display and /src/app/client)
- Add @apps path alias to tsconfig.json and vitest.config.ts
- Update all imports throughout the codebase
- Update test expectations for new route structure
- Keep shadcn/ui components in shared src/components/ui/ location
- All tests pass with 100% coverage
- Build succeeds

Resolves #13
