+++ apps/display/lib/layout-optimizer.ts
