+++ apps/display/pages/ClientPage.test.tsx
@@ -1,7 +1,7 @@
 import { describe, it, expect, beforeEach, vi } from 'vitest';
 import { render, screen, waitFor, act } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import ClientPage from './page';
+import ClientPage from './ClientPage';
 
 // Mock PeerJS
 let mockPeerHandlers: Record<string, (arg?: unknown) => void> = {};
